Move package org.candlepin.policy.js.entitlement
