+++ server/src/main/java/org/candlepin/async/tasks/EntitlerJob.java
@@ -33,7 +33,7 @@ import org.candlepin.model.dto.PoolIdAndQuantity;
 import org.candlepin.policy.EntitlementRefusedException;
 import org.candlepin.policy.ValidationError;
 import org.candlepin.policy.ValidationResult;
-import org.candlepin.policy.js.entitlement.EntitlementRulesTranslator;
+import org.candlepin.policy.entitlement.EntitlementRulesTranslator;
 
 import com.google.inject.Inject;
