+++ server/src/main/java/org/candlepin/controller/Entitler.java
@@ -40,7 +40,7 @@ import org.candlepin.model.Product;
 import org.candlepin.policy.EntitlementRefusedException;
 import org.candlepin.policy.ValidationError;
 import org.candlepin.policy.ValidationResult;
-import org.candlepin.policy.js.entitlement.EntitlementRulesTranslator;
+import org.candlepin.policy.entitlement.EntitlementRulesTranslator;
 import org.candlepin.resource.dto.AutobindData;
 import org.candlepin.service.ProductServiceAdapter;
 import org.candlepin.service.model.ContentInfo;
