+++ server/src/main/java/org/candlepin/guice/CandlepinModule.java
@@ -107,12 +107,12 @@ import org.candlepin.pki.impl.JSSPKIUtility;
 import org.candlepin.pki.impl.JSSPrivateKeyReader;
 import org.candlepin.policy.SystemPurposeComplianceRules;
 import org.candlepin.policy.criteria.CriteriaRules;
+import org.candlepin.policy.entitlement.Enforcer;
+import org.candlepin.policy.entitlement.EntitlementRules;
+import org.candlepin.policy.entitlement.EntitlementRulesTranslator;
 import org.candlepin.policy.js.JsRunner;
 import org.candlepin.policy.js.JsRunnerProvider;
 import org.candlepin.policy.js.RulesObjectMapper;
-import org.candlepin.policy.js.entitlement.Enforcer;
-import org.candlepin.policy.js.entitlement.EntitlementRules;
-import org.candlepin.policy.js.entitlement.EntitlementRulesTranslator;
 import org.candlepin.policy.js.pool.PoolRules;
 import org.candlepin.resource.ActivationKeyContentOverrideResource;
 import org.candlepin.resource.ActivationKeyResource;
