+++ server/src/main/java/org/candlepin/policy/entitlement/PreUnbindHelper.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.policy.js.entitlement;
+package org.candlepin.policy.entitlement;
 
 import org.candlepin.model.ConsumerCurator;
 import org.candlepin.policy.ValidationResult;
