+++ server/src/test/java/org/candlepin/policy/entitlement/EntitlementRulesTestFixture.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.policy.js.entitlement;
+package org.candlepin.policy.entitlement;
 
 import org.candlepin.common.config.Configuration;
 import org.candlepin.config.ConfigProperties;
