+++ server/src/test/java/org/candlepin/resource/ConsumerResourceEntitlementRulesTest.java
@@ -25,8 +25,8 @@ import org.candlepin.model.OwnerCurator;
 import org.candlepin.model.Pool;
 import org.candlepin.model.PoolCurator;
 import org.candlepin.model.Product;
-import org.candlepin.policy.js.entitlement.Enforcer;
-import org.candlepin.policy.js.entitlement.EntitlementRules;
+import org.candlepin.policy.entitlement.Enforcer;
+import org.candlepin.policy.entitlement.EntitlementRules;
 import org.candlepin.test.DatabaseTestFixture;
 import org.candlepin.test.TestDateUtil;
 import org.candlepin.test.TestUtil;
