+++ server/src/test/java/org/candlepin/resource/EntitlementResourceTest.java
@@ -35,8 +35,8 @@ import org.candlepin.model.Pool;
 import org.candlepin.model.ProductCurator;
 import org.candlepin.model.SourceStack;
 import org.candlepin.model.SubscriptionsCertificate;
-import org.candlepin.policy.js.entitlement.EntitlementRules;
-import org.candlepin.policy.js.entitlement.EntitlementRulesTranslator;
+import org.candlepin.policy.entitlement.EntitlementRules;
+import org.candlepin.policy.entitlement.EntitlementRulesTranslator;
 import org.candlepin.service.ProductServiceAdapter;
 import org.candlepin.test.TestUtil;
 import org.candlepin.util.Util;
